# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Help Center",
        "description": "Documentation for Help Center - Hackaton 2021 (winter)",
        "contact": {
            "email": "mail@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get Project",
                "description": "Get specific Project",
                "operationId": "api.projects.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "Update Project",
                "description": "Update existing Project",
                "operationId": "api.projects.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Updated project identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Project payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Project name",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 1,
                                        "example": "My Project Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Delete Project by identifier",
                "description": "Delete Project by identifier",
                "operationId": "api.projects.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfull operation"
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tutorials/{tutorialId}/steps": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Returns list of availble steps for a project",
                "description": "Returns list of availble steps for a project",
                "operationId": "api.steps.index",
                "parameters": [
                    {
                        "name": "tutorialId",
                        "in": "path",
                        "description": "Id of Tutorial",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApiResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Step"
                ],
                "summary": "Store new Step for a Tutorial",
                "description": "Store new Step for a Tutorial",
                "operationId": "api.steps.store",
                "parameters": [
                    {
                        "name": "tutorialId",
                        "in": "path",
                        "description": "Id of Tutorial",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Step payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Step title",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 1,
                                        "example": "My New Step"
                                    },
                                    "content": {
                                        "description": "Step content",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "heading-one"
                                                },
                                                "children": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "text": {
                                                                "type": "string",
                                                                "example": "Article title"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "example": "heading-one"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "type": "heading-one",
                                                "children": [
                                                    {
                                                        "text": "Article title"
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/steps/{id}": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Get Step",
                "description": "Get specific Step",
                "operationId": "api.steps.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Step",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Step"
                ],
                "summary": "Delete Step by identifier",
                "description": "Delete Step by identifier",
                "operationId": "api.steps.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Step identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfull operation"
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Step"
                ],
                "summary": "Update Step",
                "description": "Update existing Step",
                "operationId": "api.steps.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Updated Step identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Step payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Step title",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 1,
                                        "example": "My New Step"
                                    },
                                    "content": {
                                        "description": "Step content",
                                        "type": "string",
                                        "example": [
                                            {
                                                "type": "heading-one",
                                                "children": [
                                                    {
                                                        "text": "Article title"
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Api response",
                "description": "Common Api response",
                "properties": {
                    "fieldName": {
                        "title": "fieldName",
                        "description": "fieldName",
                        "type": "string",
                        "example": "field value"
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        []
    ]
}